# [Fix] `no-restricted-paths`: enhance performance for exceptions